// mobile/src/api/generated/models/getReloads200ReloadsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ReloadCars API
 * API do app de pontos de recarga para veículos elétricos.
 * OpenAPI spec version: 1.0.0
 */

export type GetReloads200ReloadsItem = {
  id: string
  /** @nullable */
  chargePointId: string | null
  chargePointName: string
  /** @nullable */
  chargePointAddress: string | null
  chargedAt: string
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  durationMinutes: number
  energyKwh: number
  energyEstimated: boolean
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   * @nullable
   */
  pricePerKwhCents: number | null
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   * @nullable
   */
  totalCents: number | null
}